requirement.txt
pymysql
sqlclient
pillow
bootstrap4
django_crispy_forms